refactor: package-priavte 적용
- package-priavte 적용
- 이게 맞는가 ?
