+++ src/main/java/com/prgrms/boardapp/service/Converter.java
@@ -1,4 +1,4 @@
-package com.prgrms.boardapp.converter;
+package com.prgrms.boardapp.service;
 
 import com.prgrms.boardapp.dto.PostDto;
 import com.prgrms.boardapp.dto.UserDto;
@@ -6,13 +6,9 @@ import com.prgrms.boardapp.model.Post;
 import com.prgrms.boardapp.model.User;
 import org.springframework.stereotype.Component;
 
-import java.util.ArrayList;
-import java.util.List;
-
 @Component
 public class Converter {
-
-    public Post convertPost(PostDto postDto) {
+    Post convertPost(PostDto postDto) {
         Post post = Post.builder()
                 .id(postDto.getId())
                 .content(postDto.getContent())
@@ -22,7 +18,7 @@ public class Converter {
         return post;
     }
 
-    public User convertUser(UserDto userDto) {
+    User convertUser(UserDto userDto) {
         return User.builder()
                 .id(userDto.getId())
                 .age(userDto.getAge())
@@ -31,7 +27,7 @@ public class Converter {
                 .build();
     }
 
-    public PostDto convertPostDto(Post post) {
+    PostDto convertPostDto(Post post) {
         return PostDto.builder()
                 .id(post.getId())
                 .createdAt(post.getCommonEmbeddable().getCreatedAt())
@@ -41,20 +37,12 @@ public class Converter {
                 .build();
     }
 
-    public UserDto convertUserDto(User user) {
+    UserDto convertUserDto(User user) {
         return UserDto.builder()
                 .id(user.getId())
                 .age(user.getAge())
                 .hobby(user.getHobby())
                 .name(user.getName())
                 .build();
-    }
-
-    public List<PostDto> convertPostDtoList(User user) {
-        List<PostDto> postDtoList = new ArrayList<>();
-        for (int i = 0; i < user.getPosts().size(); i++) {
-            postDtoList.add(this.convertPostDto(user.getPosts().get(i)));
-        }
-        return postDtoList;
     }
 }
